+++ crates/miden-testing/src/kernel_tests/tx/test_tx.rs
@@ -82,7 +82,6 @@ fn test_future_input_note_fails() -> anyhow::Result<()> {
             ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE.try_into().unwrap(),
             &[],
             NoteType::Private,
-            None,
         )
         .unwrap();
     mock_chain.prove_next_block();