+++ crates/miden-testing/tests/integration/scripts/mod.rs
@@ -1,6 +1,5 @@
 mod faucet;
 mod p2id;
 mod p2ide;
-mod p2idr;
 mod send_note;
 mod swap;